% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rudnicki.R
\name{biot_willis_effstress}
\alias{biot_willis_effstress}
\alias{biot_willis_effstress_vol}
\alias{biot_compressibility}
\alias{darcy_conductivity}
\alias{lame_first}
\title{Poroelastic parameters}
\usage{
biot_willis_effstress(nu, nuu, B)

biot_willis_effstress_vol(K, Ks)

biot_compressibility(nu, nuu, B, mu = 1)

darcy_conductivity(nu, nuu, B, mu = 1, diffusiv = 1)

lame_first(nu, mu = 1)
}
\arguments{
\item{nu}{numeric; the drained Poisson's ratio  [0,0.5]}

\item{nuu}{numeric; the undrained Poisson's ratio  [0,0.5] (\code{nu} < \code{nuu})}

\item{B}{numeric; Skempton's coefficient [0,1]}

\item{mu}{numeric; the isotropic elastic shear modulus (keep at unity for 'normalized')}

\item{diffusiv}{numeric; the isotropic hydraulic diffusivity (keep at unity for 'normalized')}
}
\description{
Poroelastic parameters
}
