% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial.R
\name{cart2pol}
\alias{cart2pol}
\title{Transform cartesian coordinates to polar}
\usage{
cart2pol(x, y, z = NULL, dist.type = c("euclidean", "manhattan"),
  radians = TRUE)
}
\arguments{
\item{x}{numeric vector, or numeric matrix (see \code{y})}

\item{y}{numeric vector; if missing, then \code{x} must have at least
two columns: the first to represent \code{x} and the second to represent \code{y}}

\item{z}{numeric vector}

\item{dist.type}{character; the type of distance calculation}

\item{radians}{logical; should the output angle be in radians? (\code{FALSE} returns in degrees)}
}
\description{
Transform cartesian coordinates to polar
}
\examples{
cart2pol(1:10, 1:10)
all.equal(cart2pol(1:10, 1:10), cart2pol(cbind(1:10, 1:10)))

# Change radius computation
cart2pol(1:10, 1:10, 1:10) # assumes 'euclidean'
cart2pol(1:10, 1:10, 1:10, 'manhattan')

# use degrees
cart2pol(1:10, 1:10, radians=FALSE)
}
\author{
modified from function in beadarrayMSV
}

