% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_range}
\alias{check_range}
\title{Check whether a quantity is wihin an acceptable range}
\usage{
check_range(x, xmin = 0, xmax = 1)
}
\arguments{
\item{x}{numeric, or coercible to numeric}

\item{xmin}{numeric; the minimum value of the acceptable range}

\item{xmax}{numeric; the maximum value of the acceptable range}
}
\value{
see \code{\link{stopifnot}}
}
\description{
Check whether a quantity is wihin an acceptable range
}
\note{
This function throws and error if range of \code{x} is outside [\code{xmin},\code{xmax}].
}
\author{
Andrew J. Barbour <andy.barbour@gmail.com>
}
