% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seismicity.R
\name{deform_hull}
\alias{deform_hull}
\title{Deform a convex hull}
\usage{
deform_hull(hull.coords, out.by = 1, shear.by = 0, plot = FALSE)
}
\arguments{
\item{hull.coords}{a matrix or data.frame with x and y coordinates of the convex hull}

\item{out.by}{numeric; the factor to expand the area of the hull by}

\item{shear.by}{numeric; the factor to shear the area of the hull by}

\item{plot}{logical; should the results be plotted?}
}
\description{
Deform the coordinates of a convex hull
}
\details{
This effectively applies a 2D plane-strain tensor to the coordinates where the 
diagonal components equal \code{out.by} and the off diagonal components equal \code{shear.by}.
}
\examples{
set.seed(1234)
X <- matrix(stats::rnorm(2000), ncol = 2)
hull.inds <- chull(X)
hull.inds <- c(hull.inds, hull.inds[1])
hull <- X[hull.inds, ]

deform_hull(hull, plot=TRUE)
deform_hull(hull, out.by=2, plot=TRUE)

}
\author{
A.J. Barbour
}
