% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreelen.R
\name{degreelen}
\alias{degreelen}
\alias{ellipsoid}
\title{Compute the length of 1-degree of latitude and longitude}
\usage{
degreelen(lat, model)

ellipsoid(model = c("IRTF", "WGS84"))
}
\arguments{
\item{lat}{numeric; degrees}

\item{model}{(optional) character; which ellipsoidal correction should be used?}
}
\description{
Compute the length of 1-degree of latitude and longitude
}
\examples{
\dontrun{
lats <- 30:32
# Lon should be about 111.4 km, and Lat 96 - 94 km
degreelen(lats)

# In most cases there is no discerible effect
# caused by varying the semi-major axis.
#   Not much difference according to this machine:
all.equal(degreelen(lats,"IRTF"), degreelen(lats,"WGS84"), tolerance = .Machine$double.eps)

#   unless the precision of the input is in micrometers, 
#   which is essentially impossible to achieve:
options(digits=13)
lats <- 30.1111111123
rbind(degreelen(lats,"IRTF"),degreelen(lats,"WGS84"))
}
}
\author{
A.J.Barbour, adapted from \code{degreelen.m} by B.W.Crowell
}
\references{
\url{http://en.wikipedia.org/wiki/World_Geodetic_System}
}

