% Generated by roxygen2 (4.0.1): do not edit by hand
\name{epsg}
\alias{D_project}
\alias{EPSG}
\alias{agency.root}
\alias{epsg}
\alias{getEPSG}
\title{Functions to work with EPSG projection specifications}
\usage{
agency.root(agency = NULL)

getEPSG(ref.id = 4326, verbose = TRUE, agency = NULL)

D_project(Df, coords = names(Df)[1:2], p4, p4old, verbose = TRUE, ...)
}
\arguments{
\item{agency}{character;}

\item{ref.id}{integer;}

\item{verbose}{logical;}

\item{Df}{an object to be coerced into a \code{\link{data.frame}}}

\item{coords}{character; the names of the coordinates in \code{Df}}

\item{p4,p4old;}{these are the proj4 strings for the new (\code{p4})
and old (\code{p4old}) data.  More specifically, \code{p4} will be
the new projection (set by \code{\link{CRS}}), and
\code{p4old} is the old projection (set by \code{\link{proj4string}}).}

\item{...}{additional parameters to \code{\link{elide}}}
}
\description{
Functions to work with EPSG projection specifications
}
\examples{
\dontrun{
library(maps)
library(maptools)
library(datasets)
data(state)
states <- data.frame(state.x77, state.center)
# HI/AK are in the middle of the ocean, so take them out:
states <- states[states$x > -121,]

# uses getEPSG to project if p4/p4old are missing
d1 <- D_project(states, c("x","y"))

# set p4
p <- getEPSG()
d2 <- D_project(states, c("x","y"), p)
all.equal(d1,d2) # TRUE

# reproject into UTM-11 in kilometers
d2u <- spTransform(d2, CRS("+proj=utm +zone=11 +ellps=WGS84 +units=km"))

# change the projection
pold <- p
p <- getEPSG(2153)  # UTM-11 in meters
d3 <- D_project(states, c("x","y"), p, pold)

layout(matrix(1:6,ncol=3))
us <- map("usa",plot=FALSE)
plot(d1, axes=TRUE); title("d1"); lines(us)
plot(d2, axes=TRUE);title("d2 (== d1)"); lines(us)
plot(d2u, axes=TRUE);title("d2 - utm11/km")
plot(d3, axes=TRUE);title("d3 - utm11/m")
#
# rotate with maptools (can also do this using ...)
plot(elide(d2u, rotate=45));title("d2 at 45") # clockwise
plot(elide(d2u, rotate=-45));title("d2 at -45") # counter-clockwise
#
}
}
\references{
A list of EPSG projections:
  \url{http://spatialreference.org/ref/epsg/}

A list of PROJ commands:
  \url{https://trac.osgeo.org/proj/wiki/GenParms}
}

