% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{erf}
\alias{erf}
\alias{erfc}
\alias{ierf}
\alias{ierfc}
\alias{ierfc2}
\title{Error functions}
\usage{
erf(x)

erfc(x)

ierf(x)

ierfc(x)

ierfc2(x)
}
\arguments{
\item{x}{numeric}
}
\description{
Error functions
}
