% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{erf_re}
\alias{erf_re}
\alias{erfc_re}
\alias{ierf_re}
\alias{ierfc_re}
\alias{ierfc2_re}
\title{Error functions}
\usage{
erf_re(x)

erfc_re(x)

ierf_re(x)

ierfc_re(x)

ierfc2_re(x)
}
\arguments{
\item{x}{numeric}
}
\description{
Error functions
}
\note{
These remain for posterity; we now use RcppFaddeeva's implementation (\code{\link[RcppFaddeeva]{erf}}, etc.).
}
