% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricecleary76.R
\name{rc76_edge_disloc_poro}
\alias{rc76_edge_disloc_poro}
\title{Rice and Cleary solutions}
\usage{
rc76_edge_disloc_poro(
  Times,
  slip,
  radial_dist,
  flt_angle = 30,
  diffusiv = 1,
  nuu = 0.33,
  nu = 0.25,
  B = 0.6,
  mu = 3e+10,
  ...
)
}
\arguments{
\item{Times}{numeric; times to calculate at \code{depth} [s]}

\item{slip}{numeric; relative slip along the edge dislocation [m]}

\item{radial_dist}{numeric; distance from xxx to xxx [m]}

\item{flt_angle}{numeric; angle counter clockwise from fault strike}

\item{diffusiv}{numeric; in plane hydraulic diffusivity [m^2/s]}

\item{nuu}{numeric; undrained Poisson's ratio [-]}

\item{nu}{numeric; drained Poisson's ratio [-]}

\item{B}{numeric; Skempton's coefficient [-]}

\item{mu}{numeric; elastic shear modulus [Pa]}

\item{...}{additional arguments}
}
\value{
tibble
}
\description{
Rice and Cleary solutions
}
\details{
\code{\link{rc76_edge_disloc_poro}} is the 2D solution
to an edge dislocation (e.g., fault) embedded in a poroelastic medium
}
\examples{
NULL
}
