% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rudnicki.R
\name{rudnicki}
\alias{rudnicki}
\alias{plot.rudnicki}
\title{Rudnicki's response to point and step injection}
\usage{
rudnicki(r, z, t, zinj = 0, mu = 1, diffusiv = 1, nu = 0.25,
  nuu = 0.33, B = 1, response = c("impulse", "step"))

\method{plot}{rudnicki}(x, response = FALSE, ...)
}
\arguments{
\item{r, z}{numeric; receiver position (radial, depth-positive) [m]}

\item{t}{numeric; time [s]}

\item{zinj}{numeric; injection depth [m]}

\item{mu}{numeric; the isotropic elastic shear modulus (keep at unity for 'normalized')}

\item{diffusiv}{numeric; the isotropic hydraulic diffusivity (keep at unity for 'normalized')}

\item{nu}{numeric; the drained Poisson's ratio  [0,0.5]}

\item{nuu}{numeric; the undrained Poisson's ratio  [0,0.5] (\code{nu} < \code{nuu})}

\item{B}{numeric; Skempton's coefficient [0,1]}

\item{response}{character; the type of reponse be the impulsive or heaviside}
}
\description{
Rudnicki's response to point and step injection
}
\examples{
r <- 1
zi <- 1
t <- seq(1,1000,length.out=101)

# Impulse response
rt <- rudnicki(r,1,t,zi)
rtdeep <- rudnicki(r,10,t,zi)

# Step response
rt_s <- rudnicki(r,1,t,zi, response='step')
rtdeep_s <- rudnicki(r,10,t,zi, response='step')

layout(matrix(1:6,2))
plot(rt, col=1)
plot.new()

plot(rtdeep, col=2)
plot.new()

plot(rt_s, col=3)
plot.new()

plot(rtdeep_s, col=4)
plot.new()

}
