% Generated by roxygen2 (4.0.1): do not edit by hand
\name{segall85}
\alias{.surface_g}
\alias{segall85}
\alias{surface_displacement}
\alias{timevarying_surface_displacement}
\title{Surface deformation associated with fluid withdrawl}
\usage{
segall85()

surface_displacement(y, C. = 1, mu. = 1e+09, ...)

.surface_g(y = 0, z_src = 0, y_src = 0, nuu = 1/3)

timevarying_surface_displacement(y, Time, Vdot., B., L., D., HD., nuu. = 1/3,
  y_src. = 0)
}
\description{
Surface deformation associated with fluid withdrawl
}
\examples{
\dontrun{
y. <- c(-10:-2, seq(-1.9,1.9,by=0.1), 2:10)
su <- surface_displacement(y.*1e3, z_src=1e3)
plot(uz ~ y, su, type="b")
}
}

