% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segall89.R
\name{segall89}
\alias{.surface_g_xi}
\alias{segall1989}
\alias{segall89}
\alias{surface_displacement_reservoir}
\title{Surface deformation associated with fluid withdrawl from a disk-shaped reservoir}
\usage{
segall89(help = FALSE)

.surface_g_xi(x, depth, halfwidth, thickness)

surface_displacement_reservoir(x, depth = 1, halfwidth = depth,
  thickness = depth/10, DelM. = 1, nuu. = 1/3, B. = 1, rho_f. = 1000,
  ...)
}
\arguments{
\item{help}{logical; load documentation for \code{\link{segall89}}}

\item{x}{numeric; the horizontal (or radial) distance from the source}

\item{depth}{numeric; the depth below the surface to the top of the reservoir}

\item{halfwidth}{numeric; the half-length of the reservoir}

\item{thickness}{numeric; the thickness of the reservoir}

\item{DelM.}{numeric; the change in mass per unit volume in the reservoir}

\item{nuu.}{numeric; the undrained Poisson's ratio for the reservoir material}

\item{B.}{numeric; Skempton's ratio for the reservoir material}

\item{rho_f.}{numeric; the density of the fluid in the reservoir material}

\item{...}{additional parameters (currently unused)}
}
\description{
Surface deformation associated with fluid withdrawl from a disk-shaped reservoir
}
\examples{
# Reproduce Figure 5 from Segall (1989)
r <- seq(-4,4,by=0.1)
xx <- surface_displacement_reservoir(r)
plot(Xi_z ~ x, xx, type='l', ylim=c(-1.6,0.8))
lines(Xi_x ~ x, xx)
lines(Xi_xx ~ x, xx)
}
\seealso{
\code{\link{segall85}}; \code{\link{Simple-deformation}}
}

