% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segall92.R
\name{segall92}
\alias{segall92}
\alias{surface_displacement_ringdisk}
\title{Elastostatic subsidence of a half space}
\usage{
segall92()

surface_displacement_ringdisk(
  x,
  radius,
  depth,
  thickness,
  p0 = 1,
  nuu = 1/4,
  ShearModulus = 1.5e+10,
  BiotCoef = 1,
  verbose = TRUE,
  x.lim,
  tol_pow,
  nsub
)
}
\arguments{
\item{x}{}

\item{radius, depth, thickness}{numeric; the radius, midpoint depth, and thickness of the circular-disk pressure source; the
midpoint-depth must be at least \code{thickness/2}}

\item{p0}{numeric; the (uniform) pressure change inside the circular-disk source}

\item{nuu}{numeric; the undrained Poisson's ratio}

\item{ShearModulus}{numeric; the elastic shear modulus}

\item{BiotCoef}{numeric; Biot's coefficient (usually written as \eqn{\alpha});
Note that this is converted to Geertsma's coefficent by \code{BiotCoef/ShearModulus}}

\item{verbose}{logical; should messages be given?}

\item{x.lim}{numeric; the limit of integration in the radial direction}

\item{tol_pow}{numeric; the exponent applied to \code{.Machine$double.eps} used as 
the integration tolerance (see \code{\link{integrate}}); \emph{Increasing this to
increase inegration accuracy will increase computation time}}

\item{nsub}{integer; the number of subdivisions in the integration (see \code{\link{integrate}})}
}
\description{
Elastostatic subsidence of a half space
}
\examples{

# Calculate surface displacements from a pressure-disk source

# Radial distances (e.g., observation points) and pressure source strength
# (in this case, keep distances in km to stabilize integration
#  and then rescale pressure to maintain unit consistency)
r <- seq(0, 15, length.out=151)
Pa <- 80e3 #Pa (N / m^2)
pressure <- 2.5 * Pa * (1000^2)

# dimensions of inflating (pressure > 0) or deflating (pressure < 0) reservoir
disk.radius <- 4
disk.thickness <- 2
disk.middepth <- 3

# calculations...
res <- surface_displacement_ringdisk(r, radius = disk.radius, thickness = disk.thickness, depth = disk.middepth, p0 = pressure)
head(res)

with(res,{
  uz_sc <- -uz
  ur_sc <- ur
  yl <- range(c(uz_sc, ur_sc, err), na.rm=TRUE)
  plot(x, uz_sc, col=NA, log='', ylim=yl, xaxs='i',
       ylab="Displacement, mm", xlab="Distance from source, km", main="Surface Displacements: Pressure-disk Source")
  abline(h=0, col='grey')

  lines(x, uz_sc, lwd=1.5)
  text(10, 7, "Vertical")

  lines(x, ur_sc, lty=2, lwd=1.5)
  text(5, 5, "Radial")

  lines(x, err, lty=3, lwd=1.5)
  text(4.5, -0.75, "Radial strain\n(mm/km)", pos=2)

  mtext(sprintf("Segall (1992) model with radius: \%s, Thickness: \%s, Depth: \%s", disk.radius, disk.thickness, disk.middepth), font=3)
})

}
\references{
Segall, P. (1992), Induced stresses due to fluid extraction from axisymmetric reservoirs,
PAGEOPH, 139: 535. doi: 10.1007/BF00879950
}
