% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segall85.R
\name{Simple-deformation}
\alias{.setleft}
\alias{Simple-deformation}
\alias{Tilt}
\alias{Uniaxial_extension}
\title{Simple numerical deformation estimates}
\usage{
Uniaxial_extension(x, X, left = TRUE)

Tilt(x, y = NULL, z, left = TRUE)

.setleft(x, left = TRUE)
}
\arguments{
\item{x}{numeric; the spatial vector in real units (not an index!)}

\item{X}{numeric; the quantity varying in the direction of \code{x}}

\item{left}{logical; should padding be on the "left" side of the vector (i.e., index number 1)?}

\item{y}{numeric; the spatial vector perpendicular to \code{x} in real units}

\item{z}{numeric; values which vary in the direction perpendicular to the \code{x}-\code{y} plane}
}
\description{
Calculate tilts and extensions based on spatially varying displacements
}
\details{
\code{\link{Uniaxial_extension}} calculates the component of
strain associated with deformation along a single axis.
For example, the change in the Eastward displacements (or rates)
in the East direction.

\code{\link{Tilt}} calculates the tilt field associates with
spatial variations in vertical positions (or rates of change);
the sign convention used is such that
a ball placed on the x-y plane would roll in the direction of the
tilt vector.  Or, in other words, the tilt vector is the direction a
plumb bob would move.

Calculations are done with \code{\link{diff}}.
}
\note{
\code{\link{Tilt}} has not been well-tested for two-dimensional
results!
}
\examples{
# Some x values
xval. <- sort(unique(c(-7:-3, seq(-3.90,3.90,by=0.05), 3:7)))
set.seed(1221)
xanis <- rnorm(length(xval.), sd = 10)

#  surface displacements
su <- surface_displacement(xval.*1e3, C.=1e13, z_src=0.7e3)

# Vertical tilt -- assumes axial symmetry
sut <- with(su, Tilt(x, z=uz))
#               -- including anisotropic effects
sut.anis <- with(su, Tilt(x, x = sort(x + xanis), z=uz))

plot(ztilt ~ x, sut.anis, col='blue', pch=16, cex=0.5)
lines(ztilt ~ x, sut, lwd=2)

plot(ztilt ~ abs(x), sut.anis, col='blue', pch=ifelse(sign(x)==1,16,1), cex=0.5)
lines(ztilt ~ abs(x), sut, lwd=2)
 
# Uniaxial strain in the 'x' direction
sue <- with(su, Uniaxial_extension(x, X=ux))
sue.anis <- with(su, Uniaxial_extension(sort(x + xanis), X=ux))

plot(dXdx ~ x, sue.anis, col='blue', pch=16, cex=0.5)
lines(dXdx ~ x, sue, lwd=2)
 
plot(dXdx ~ abs(x), sue.anis, col='blue', pch=ifelse(sign(x)==1,16,1), cex=0.5)
lines(dXdx ~ abs(x), sue, lwd=2)

# see how the 'left' argument affects things:
.setleft(1,TRUE) # NA  1
.setleft(1,FALSE) # 1  NA
}
\seealso{
\code{\link{segall85}}
}

