% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deform-package.R
\docType{package}
\name{deform}
\alias{deform}
\alias{deform-package}
\title{The \code{deform} package: crustal deformation tools in R}
\description{
The \code{deform} package: crustal deformation tools in R
}
\seealso{
\code{\link{segall85}} for deformation associated with fluid extraction
}
