% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deform-utilities.R
\name{erf}
\alias{erf}
\alias{erfc}
\alias{ierf}
\alias{ierfc}
\title{Error functions}
\usage{
erf(x)

erfc(x)

ierf(x)

ierfc(x)
}
\arguments{
\item{x}{numeric;}
}
\description{
Error functions
}

