% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geertsma.R
\name{geertsma73}
\alias{.surface_g_pdisk}
\alias{geertsma1973}
\alias{geertsma73}
\alias{surface_displacement_pressuredisk}
\title{Surface deformation associated with fluid withdrawl from a symmetric, rectangular reservoir}
\usage{
geertsma73(help = FALSE)

.surface_g_pdisk(x = 0, x_src = 0, z_src = 0, nuu = 1/3, mu., cm)

surface_displacement_pressuredisk(x, depth, thickness, radius, delP.,
  shear.mod = 5e+09, ...)
}
\arguments{
\item{help}{logical; load documentation for \code{\link{geertsma77}}}

\item{x}{numeric; the horizontal (or radial) distance from the source}

\item{mu.}{numeric; the shear modulus}

\item{depth}{numeric; the depth below the surface to the source}

\item{thickness}{numeric;}

\item{radius}{numeric;}

\item{delP.}{numeric; uniform change in pore-fluid pressure (positive = increase)}

\item{...}{additional arguments passed to \code{\link{.surface_g_pdisk}}}
}
\description{
Surface deformation associated with fluid withdrawl from a symmetric, rectangular reservoir
}
\examples{

# Reproduce xxx
r <- seq(-10,10,by=0.2) # km
r.m <- r*1e3
xx <- surface_displacement_pressuredisk(r.m, depth=2e3, thickness=100, radius=1e3, delP.=1e3)
plot(uz ~ x, xx, type='l')

}
\seealso{
\code{\link{vasco98}}; \code{\link{Simple-deformation}}
}

