% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hydraulic_diffusivity}
\alias{hydraulic_diffusivity}
\alias{rigid_hydraulic_diffusivity}
\title{Calculate hydraulic diffusivity of a fluid saturated medium}
\usage{
hydraulic_diffusivity(k., eta., mu., B., nu. = 1/4, nuu. = 1/3)

rigid_hydraulic_diffusivity(k., eta., phi., Beta.)
}
\description{
Calculate hydraulic diffusivity of a fluid saturated medium
}

