% Generated by roxygen2 (4.0.1): do not edit by hand
\name{magnitude-distance}
\alias{dynamic}
\alias{magnitude-distance}
\alias{static}
\title{Predict the level of strain expected for magnitude and distance}
\usage{
static(Mw, Distance.km, model = "w88")

dynamic(Mw, Distance.km, km2deg = NULL, model = "aw14")
}
\description{
Predict the level of strain expected for magnitude and distance
}

