% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_magdist_scaling.R
\name{magnitude-distance}
\alias{magnitude-distance}
\alias{static}
\alias{dynamic}
\title{Predict the level of strain expected for magnitude and distance}
\usage{
static(Mw, Distance.km, model = "w88")

dynamic(
  Mw,
  Distance.km,
  model = c("aw14", "bc17", "fbl20", "blf21"),
  strn.type = c("general", "areal", "diffext", "shear", "dil"),
  km2deg = NULL
)
}
\arguments{
\item{Mw}{numeric; the moment magnitude of the earthquake}

\item{Distance.km}{numeric; the epicentral distance, in kilometers}

\item{model}{character; the strain-scaling model to use (see Details)}

\item{strn.type}{character; the type of strain-scaling relationship to
provide a prediction for}

\item{km2deg}{numeric; specify the scaling coefficient used to
convert the distance (in km) to degrees; this may be useful if,
for example, the distance is already in degrees (then set it equal to 1)}
}
\description{
Predict the level of strain expected for magnitude and distance
}
\details{
Currently the options for \code{model} are:
\describe{
  \item{\code{'static'}}{Static RMS strain from an earthquake (Wyatt, 1988)}
  \item{\code{'dynamic'}}{Dynamic RMS strain from a seismic wave:
  \code{'aw14'} for Agnew and Wyatt (2014): applicable to teleseisms at thousands of km (inaccurate within 500km or so); or
  \code{'bc17'} for Barbour and Crowell (2017): applicable to regional earthquakes at < 500 km (inaccurate
  for very short distances); or
  \code{'fbl20'} for Farghal, Barbour and Langbein (2020): applicable to regional earthquakes 
  at < 500 km (best accuracy at short distances; inaccurate > 500 km);
  \code{'blf21'} for Barbour, Langbein, and Farghal (2021): best for regional earthquakes as
  it accounts for station and event terms in the fbl20 dataset; the basis of the M_{DS} 
  magnitude scale, which is basically equivalent to Mw.
  }
}
}
\examples{
static(7.2, 220)
dynamic(7.2, 15000, model='aw14')

dynamic(7.2, 150, model='aw14')
dynamic(7.2, 150, model='bc17')
dynamic(7.2, 150, model='fbl20')
}
\references{
Agnew, D. C., and F. K. Wyatt (2014),
Dynamic strains from regional and teleseismic earthquakes, 
Bull. Seismol. Soc. Am. 104, no. 4, 1846– 1859, 
doi: 10.1785/0120140007

Barbour, A., and B. Crowell (2017),
Dynamic strains for earthquake source characterization, 
Seismol. Res. Lett. 88, no. 2A, 354–370, 
doi: 10.1785/0220160155

Farghal, N., A. Barbour, and J. Langbein (2020),
The Potential of Using Dynamic Strains in Earthquake Early Warning Applications, 
Seismol. Res. Lett., in press, 1–12, 
doi: 10.1785/0220190385

Barbour, A., J. Langbein, and N. Farghal (2021),
Earthquake Magnitudes from Dynamic Strain,
Bull. Seismol. Soc. Am.m, 111 (3): 1325–1346
doi: 10.1785/0120200360

Wyatt, F. K. (1988),
Measurements of coseismic deformation in southern California: 1972–1982, 
J. Geophys. Res. 93, no. B7, 7923–7942, 
doi: 10.1029/JB093iB07p07923
}
