% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricecleary76.R
\name{rc76_edge_disloc_poro}
\alias{rc76_edge_disloc_poro}
\title{Rice and Cleary solutions}
\usage{
rc76_edge_disloc_poro(
  Times,
  slip,
  radial_dist,
  angFromEast = 30,
  diffusiv = 1,
  nuu = 0.33,
  nu = 0.25,
  B = 0.6,
  mu = 3e+10,
  ...
)
}
\arguments{
\item{Times}{numeric; times to calculate at \code{depth} [s]}

\item{slip}{numeric; relative slip along the edge dislocation [m]}

\item{radial_dist}{numeric; distance from xxx to xxx [m]}

\item{angFromEast}{}

\item{diffusiv}{}

\item{nuu}{}

\item{nu}{}

\item{B}{}

\item{mu}{}

\item{...}{}
}
\value{

}
\description{
Rice and Cleary solutions
}
\details{
\code{\link{rc76_edge_disloc_poro}} is the 2D solution
to an edge dislocation (e.g., fault) embedded in a poroelastic medium
}
\examples{
NULL
}
