% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rudnicki.R
\name{rudnicki}
\alias{rudnicki}
\alias{plot.rudnicki}
\title{Rudnicki's response to point and step injection}
\usage{
rudnicki(r, z, t, zinj = 0, mu = 1e+10, d = 1, nu = 0.25, nuu = 0.33,
  B = 1, response = c("impulse", "step"))

\method{plot}{rudnicki}(x, response = FALSE, ...)
}
\arguments{
\item{r, z}{numeric; receiver position (radial, depth-positive) [m]}

\item{t}{numeric; time [s]}

\item{zinj}{numeric; injection depth [m]}

\item{mu}{numeric; elastic shear modulus [Pa]}

\item{d}{numeric; hydraulic diffusivity [m^2/s]}

\item{nu, nuu}{numeric; drained and undrained Poisson's ratio [0,0.5]}

\item{B}{numeric; Skempton's coefficient [0,1]}

\item{response}{character; the type of reponse be the impulsive or heaviside}
}
\description{
Rudnicki's response to point and step injection
}
\examples{
r <- 1
zi <- 1
t <- seq(1,1000,length.out=101)
mu <- 15e9

# Impulse response
rt <- rudnicki(r,1,t,zi, mu=mu)
rtdeep <- rudnicki(r,10,t,zi, mu=mu)

# Step response
rt_s <- rudnicki(r,1,t,zi, mu=mu, response='step')
rtdeep_s <- rudnicki(r,10,t,zi, mu=mu, response='step')

layout(matrix(1:6,2))
plot(rt, col=1)
plot.new()

plot(rtdeep, col=2)
plot.new()

plot(rt_s, col=3)
plot.new()

plot(rtdeep_s, col=4)
plot.new()

}
