% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rudnicki86.R
\name{rudnicki86}
\alias{rudnicki86}
\alias{plot.rudnicki.pt}
\title{Rudnicki's response to point and step injection}
\usage{
rudnicki86(r, z, t, zinj = 0, mu = 1, diffusiv = 1, nu = 0.25,
  nuu = 0.33, B = 1, response = NULL)

\method{plot}{rudnicki.pt}(x, response = FALSE, ...)
}
\arguments{
\item{r, z}{numeric; receiver position (radial, depth-positive) [m]}

\item{t}{numeric; time [s]}

\item{zinj}{numeric; injection depth [m]}

\item{mu}{numeric; the isotropic elastic shear modulus (keep at unity for 'normalized')}

\item{diffusiv}{numeric; the isotropic hydraulic diffusivity (keep at unity for 'normalized')}

\item{nu}{numeric; the drained Poisson's ratio  [0,0.5]}

\item{nuu}{numeric; the undrained Poisson's ratio  [0,0.5] (\code{nu} < \code{nuu})}

\item{B}{numeric; Skempton's coefficient [0,1]}

\item{response}{character; the type of response to output: 
either the points-source impulsive (\code{'impulse'}) or Heaviside (\code{'step'}) response}

\item{x}{an object with class \code{'rudnicki.pt'}, e.g. from \code{\link{rudnicki86}}}

\item{...}{additional arguments}
}
\description{
Rudnicki's response to point and step injection
}
\examples{
r <- 1
zi <- 1
t <- 10^seq(-4, 1, length.out=301)

# Impulse response
rt <- rudnicki86(r,1,t,zi)
rtdeep <- rudnicki86(r,10,t,zi)

# Step response
rt_s <- rudnicki86(r,1,t,zi, response='step')
rtdeep_s <- rudnicki86(r,10,t,zi, response='step')

layout(matrix(1:10, 2, byrow=TRUE))
plot(rt, col=1)
plot(rt, response=TRUE, col=1)

plot(rtdeep, col=2)
plot(rtdeep, response=TRUE, col=2)

plot(rt_s, col=3)
plot(rt_s, response=TRUE, col=3)

plot(rtdeep_s, col=4)
plot(rtdeep_s, response=TRUE, col=4)

}
\references{
Rudnicki, J. W. (1986), Fluid mass sources and point forces in linear elastic diffusive solids, 
\emph{Mechanics of Materials}, 5(4), 383-393, \url{https://doi.org/10.1016/0167-6636(86)90042-6}
}
