% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/segall85.R
\name{segall85}
\alias{.surface_g}
\alias{segall}
\alias{segall85}
\alias{surface_displacement}
\alias{timevarying_fluidmass}
\alias{timevarying_porepressure}
\alias{timevarying_surface_displacement}
\title{Surface deformation associated with fluid withdrawl}
\usage{
segall85(help = FALSE)

surface_displacement(x, C. = 1, mu. = 1e+09, ...)

.surface_g(x = 0, x_src = 0, z_src = 0, nuu = 1/3)

timevarying_fluidmass(x, Time, Vdot., L., t., HD., phi.)

timevarying_surface_displacement(x, Time, Vdot., B., L., D., HD., nuu. = 1/3,
  Pt.Sources.x = 0, x.lim)

timevarying_porepressure(x, z, Time, Vdot., B., L., D., HD., t., mu.gpa.,
  nu. = 1/4, nuu. = 1/3, Pt.Sources.x = 0, x.lim)
}
\arguments{
\item{help}{logical; load documentation for \code{\link{segall85}}}

\item{x}{numeric; spatial coordinate relative to extraction point}

\item{C.}{numeric; units of force, proportional to source strength (e.g., extraction rate)}

\item{mu.}{numeric; the shear modulus in Pascals}

\item{...}{additional arguments to \code{\link{.surface_g}}}

\item{x_src}{numeric; the horizontal distance from the source}

\item{z_src}{numeric; the depth of the source below the surface}

\item{nuu}{numeric; the 'undrained' Poisson's ratio (typically 1/3)}

\item{Time}{numeric; the time from xxx}

\item{Vdot.}{numeric; the volumetric flow rate xxx}

\item{L.}{numeric; the xxx}

\item{t.}{numeric; the xxx}

\item{HD.}{numeric; the xxx}

\item{phi.}{numeric; the xxx}

\item{B.}{numeric; the xxx}

\item{D.}{numeric; the xxx}

\item{Pt.Sources.x}{numeric; a vector of point-source locations in the x direction}

\item{x.lim}{numeric; the limit of integration in both the positive and negative directions; if
missing this is based on the absolute maximum of \code{x}}

\item{mu.gpa.}{numeric; the shear modulus in giga-Pascals (GPa)}

\item{nu.}{numeric; the drained Poisson's ratio (typically 1/4)}
}
\description{
Surface deformation associated with fluid withdrawl
}
\examples{
\dontrun{
x. <- c(-10:-2, seq(-1.9,1.9,by=0.1), 2:10)
su <- surface_displacement(x.*1e3, z_src=1e3)
plot(uz ~ x, su, type="b")
}
}
\seealso{
\code{\link{Simple-deformation}}
}

