% Generated by roxygen2 (4.0.1): do not edit by hand
\name{segall85}
\alias{.surface_g}
\alias{segall85}
\alias{surface_displacement}
\alias{timevarying_fluidmass}
\alias{timevarying_porepressure}
\alias{timevarying_surface_displacement}
\title{Surface deformation associated with fluid withdrawl}
\usage{
segall85()

surface_displacement(x, C. = 1, mu. = 1e+09, ...)

.surface_g(x = 0, x_src = 0, z_src = 0, nuu = 1/3)

timevarying_fluidmass(x, Time, Vdot., L., t., HD., phi.)

timevarying_surface_displacement(x, Time, Vdot., B., L., D., HD., nuu. = 1/3,
  Pt.Sources.x = 0, x.lim = 2 * round(max(abs(x), na.rm = TRUE)))

timevarying_porepressure(x, z, Time, Vdot., B., L., D., HD., t., mu.gpa.,
  nu. = 1/4, nuu. = 1/3, Pt.Sources.x = 0, x.lim = round(max(abs(x),
  na.rm = TRUE))/1)
}
\description{
Surface deformation associated with fluid withdrawl
}
\examples{
\dontrun{
x. <- c(-10:-2, seq(-1.9,1.9,by=0.1), 2:10)
su <- surface_displacement(x.*1e3, z_src=1e3)
plot(uz ~ x, su, type="b")
}
}
\seealso{
\code{\link{Simple-deformation}}
}

