% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Simple-deformation}
\alias{.setleft}
\alias{Simple-deformation}
\alias{Tilt}
\alias{Uniaxial_extension}
\title{Simple numerical deformation estimates}
\usage{
Uniaxial_extension(x, X, left = TRUE)

Tilt(x, y = NULL, z, left = TRUE)

.setleft(x, left = TRUE)
}
\arguments{
\item{x}{numeric; the spatial vector in real units (not an index!)}

\item{X}{numeric; the quantity varying in the direction of \code{x}}

\item{y}{numeric; the spatial vector perpendicular to \code{x} in real units}

\item{z}{numeric; values which vary in the direction perpendicular to the \code{x}-\code{y} plane}

\item{left}{logical; should padding be on the "left" side of the vector (i.e., index number 1)?}
}
\description{
Simple numerical deformation estimates
}
\details{
\code{\link{Uniaxial_extension}} calculates the component of
strain associated with deformation along a single axis.
For example, the change in the Eastward displacements (or rates)
in the East direction.

\code{\link{Tilt}} calculates the tilt field associates with
spatial variations in vertical positions (or rates of change);
the sign convention used is such that
a ball placed on the x-y plane would roll in the direction of the
tilt vector.  Or, in other words, the tilt vector is the direction a
plumb bob would move.
}
\note{
\code{\link{Tilt}} has not been well-tested for two-dimensional
results!
}
\examples{
# Some x values
xval. <- sort(unique(c(-7:-3, seq(-3.90,3.90,by=0.15), 3:7)))

#  surface displacements
su <- surface_displacement(xval.*1e3, C.=1e13, z_src=0.7e3)

# Vertical tilt -- assumes axial symmetry
sut <- with(su, Tilt(x, z=uz))
# including anisotropic effects
sut <- with(su, Tilt(x, x=x+rnorm(length(x)), z=uz))

# Calculate strain in the 'x' direction
sue <- with(su, Uniaxial_extension(x, X=ux))

# see how the 'left' argument affects things:
.setleft(1,TRUE) # NA  1
.setleft(1,FALSE) # 1  NA
}
\seealso{
\code{\link{segall85}}
}

